+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Breaking
 
 ### Features
+* Headers from incoming http requests can now be copied into events via `copy_headers_to_log` config in http input, `collect_meta` will be deprecated in the future
 
 ### Improvements